Move contexts to separate files
