+++ mcmd/commands/import_.py
@@ -121,7 +121,7 @@ def _import_from_url(args):
 
 def _import_from_quick_folders(args):
     file_name = os_path.splitext(args.resource)[0]
-    file_map = scan_folders_for_files(config.git_paths() + _get_dataset_folders())
+    file_map = scan_folders_for_files(config.git_paths() + context().get_dataset_folders())
     path = select_path(file_map, file_name)
     _do_import(path, args.to_package, args.entity_type_id)
 
@@ -248,7 +248,3 @@ def _poll_for_completion(url):
                  poll_forever=True)
     import_run = get(url).json()
     return import_run['status'], import_run['message']
-
-
-def _get_dataset_folders():
-    return [Path(folder) for folder in config.get('resources', 'dataset_folders')]
